api/{{ cookiecutter.api_name }}/requirements.txt

#git+https://bitbucket.sulamerica.com.br/scm/an/sas-medcon-arco-model.git

python-dotenv
pytest
graypy==2.1.0
joblib==0.16.0
PyYAML==5.3.1
fastapi==0.63.0
uvicorn==0.13.3
locust==1.4.1
httptools==0.1.1
gunicorn==20.0.4
uvloop==0.14.0

